Update grad surg batch size
Grad surg batch size was set to be a half of correct number, resulting in only half of the paraphrase data were trained. Update grad_surg.py to fix that. The prediction folder contains result trained with only half data. This training used 5e-5 learning rate, and is too large.